backend/requirements.txt
# FastAPI e dependências principais
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Banco de dados
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Armazenamento de arquivos (MinIO/S3)
boto3==1.34.0
minio==7.2.0

# HTTP client para LangFlow
httpx==0.25.2
aiohttp==3.9.1

# Processamento de arquivos
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-magic==0.4.27

# Logging
loguru==0.7.2

# Utilitários
python-dotenv==1.0.0
typing-extensions==4.8.0
email-validator==2.1.0
python-dateutil==2.8.2

# Validação de dados brasileiros
python-cpf-cnpj==1.0.0

# Timezone
pytz==2023.3

# JSON e serialização
orjson==3.9.10

# Desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0
faker==20.1.0

# Formatação e linting
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoramento
prometheus-client==0.19.0

# Cache
redis==5.0.1
aioredis==2.0.1